Another optimization: short circuit visibility logic if we don't care if the space object is scanned or just visible
