+++ FrEee/Game/Objects/Vehicles/Satellite.cs
@@ -1,6 +1,7 @@
 ﻿using FrEee.Game.Enumerations;
 using FrEee.Game.Interfaces;
 using FrEee.Game.Objects.Civilization;
+using FrEee.Utility.Extensions;
 using System;
 
 namespace FrEee.Game.Objects.Vehicles
@@ -47,7 +48,7 @@ namespace FrEee.Game.Objects.Vehicles
 		{
 			var vis = base.CheckVisibility(emp);
 			var sobj = Container as ISpaceObject;
-			if (vis < Visibility.Scanned && sobj != null && sobj.CheckVisibility(emp) >= Visibility.Scanned)
+			if (vis < Visibility.Scanned && sobj != null && sobj.HasVisibility(emp, Visibility.Scanned))
 				vis = Visibility.Scanned;
 			return vis;
 		}
