+++ FrEee/Utility/Extensions/CommonExtensions.cs
@@ -1786,69 +1786,6 @@ namespace FrEee.Utility.Extensions
 			return "The " + emp.Name;
 		}
 
-		internal static Visibility CheckSpaceObjectVisibility(this ISpaceObject sobj, Empire emp)
-		{
-			bool hasMemory = false;
-			if (sobj.IsMemory)
-			{
-				var mowner = sobj.MemoryOwner();
-				if (mowner == emp || mowner == null)
-					return Visibility.Fogged;
-				else
-					return Visibility.Unknown; // can't see other players' memories
-			}
-			else
-			{
-				var mem = sobj.FindMemory(emp);
-				if (mem != null)
-					hasMemory = true;
-			}
-
-			if (emp == sobj.Owner)
-				return Visibility.Owned;
-
-			// You can always scan space objects you are in combat with.
-			// But only their state at the time they were in combat; not for the rest of the turn!
-			// TODO - what about glassed planets, they have no owner...
-			if (Galaxy.Current.Battles.Any(b =>
-			(b.Combatants.OfType<ISpaceObject>().Contains(sobj)
-				|| b.StartCombatants.Values.OfType<ISpaceObject>().Contains(sobj)
-				|| b.EndCombatants.Values.OfType<ISpaceObject>().Contains(sobj))
-			&& b.Combatants.Any(c => c.Owner == emp)))
-				return Visibility.Scanned;
-
-			// do we have anything that can see it?
-			var sys = sobj.StarSystem;
-			if (sys == null)
-				return Visibility.Unknown;
-			var seers = sys.FindSpaceObjects<ISpaceObject>(s => s.Owner == emp && !s.IsMemory);
-			if (!seers.Any() || sobj.IsHiddenFrom(emp))
-			{
-				if (Galaxy.Current.OmniscientView && sobj.StarSystem.ExploredByEmpires.Contains(emp))
-					return Visibility.Visible;
-				if (emp.AllSystemsExploredFromStart)
-					return Visibility.Fogged;
-				var known = emp.Memory[sobj.ID];
-				if (known != null && sobj.GetType() == known.GetType())
-					return Visibility.Fogged;
-				else if (Galaxy.Current.Battles.Any(b => b.Combatants.Any(c => c.ID == sobj.ID) && b.Combatants.Any(c => c.Owner == emp)))
-					return Visibility.Fogged;
-				else if (hasMemory)
-					return Visibility.Fogged;
-				else
-					return Visibility.Unknown;
-			}
-			if (!sobj.HasAbility("Scanner Jammer"))
-			{
-				var scanners = seers.Where(s =>
-					s.HasAbility("Long Range Scanner") && s.GetAbilityValue("Long Range Scanner").ToInt() >= s.Sector.Coordinates.EightWayDistance(sobj.FindSector().Coordinates)
-					|| s.HasAbility("Long Range Scanner - System"));
-				if (scanners.Any())
-					return Visibility.Scanned;
-			}
-			return Visibility.Visible;
-		}
-
 		private static Expression CreateParam(ParameterExpression[] paramsOfDelegate, int i, ParameterInfo callParamType, Queue<object> queueMissingParams)
 		{
 			if (i < paramsOfDelegate.Length)
